+++ code/theotherbook/3.7/11.cpp
@@ -4,8 +4,8 @@
 void main() {
     double S=0, sum=0;
     
-    for(double i=1; i<=20; i++) {
-        for(double j=1; j<=30; j++) {
+    for(int i=1; i<=20; i++) {
+        for(int j=1; j<=30; j++) {
             sum += (i+j)/(i*j+3);
         }
         S += sum;
