+++ code/theotherbook/3.7/14.cpp
@@ -4,8 +4,8 @@
 void main() {
     double S=0, prod=1;
     
-    for(double i=10; i<=20; i++) {
-        for(double j=1; j<=i; j++) {
+    for(int i=10; i<=20; i++) {
+        for(int j=1; j<=i; j++) {
             prod *= sin(i)+cos(j);
         }
         S += prod;
