+++ code/theotherbook/3.7/18.cpp
@@ -4,8 +4,8 @@
 void main() {
     double W=1, sum=0;
     
-    for(double m=0; m<=6; m++) {
-        for(double n=1; n<=5; n++) {
+    for(int m=0; m<=6; m++) {
+        for(int n=1; n<=5; n++) {
             sum+=(n+m)/2;
         }
         W*=sum;
