+++ code/theotherbook/3.7/19.cpp
@@ -4,8 +4,8 @@
 void main() {
     double S=0, sum=0;
     
-    for(double i=0; i<=6; i++) {
-        for(double j=1; j<=5; j++) {
+    for(int i=0; i<=6; i++) {
+        for(int j=1; j<=5; j++) {
             sum+=cos(j*j)+i;
         }
         S+=pow(sin(i+1),2)+sum;
