+++ code/theotherbook/3.7/4.cpp
@@ -3,8 +3,8 @@
 
 void main() {
     double P=1, prod=1;
-    for(double i=1; i<=15; i++) {
-        for(double m=1; m<=6; m++) {
+    for(int i=1; i<=15; i++) {
+        for(int m=1; m<=6; m++) {
             prod *= i/(i+m);
         }
         P *= prod;
