+++ code/theotherbook/3.7/8.cpp
@@ -4,8 +4,8 @@
 void main() {
     double S=0, prod=1;
     
-    for(double i=1; i<=10; i++) {
-        for(double j=1; j<=7; j++) {
+    for(int i=1; i<=10; i++) {
+        for(int j=1; j<=7; j++) {
             prod *= (i+j)/(i*j);
         }
         S += prod;
