+++ code/theotherbook/3.7/9.cpp
@@ -4,8 +4,8 @@
 void main() {
     double S=0, sum=0;
     
-    for(double i=1; i<=10; i++) {
-        for(double j=1; j<=8; j++) {
+    for(int i=1; i<=10; i++) {
+        for(int j=1; j<=8; j++) {
             sum += sqrt(pow(i,2)+pow(j,2));
         }
         S += pow(i,2) + sum;
